# api_gateway/docs/swagger.yaml
basePath: /
info:
  contact: {}
  title: kpokr2 API
paths:
  /analyze/{id}:
    get:
      description: Analyzes a previously uploaded text file and returns statistics
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Analyze file by ID
  /files/{id}:
    get:
      description: Retrieves the content of a previously uploaded text file
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get file content by ID
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a text file and returns a unique file ID
      parameters:
      - description: Text file to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload a text file
swagger: "2.0"